Added the functionality to DataDisplay
I made a mistake and added the functionality to home so I put it in DataDisplay instead
