+++ my-app/public/words.txt
@@ -14,3 +14,65 @@ Moving the goal post
 ASAP
 Paradigm shift
 Yield
+All hands on deck
+At the end of the day
+Back to the drawing board
+Backburner
+Bite the bullet
+Bleeding edge
+Blue sky thinking
+Boil the ocean
+Break down the silos
+Bring to the table
+Buy-in
+Cascade information
+Circle back
+Close of play
+Collaborate
+Core competencies
+Deep dive
+Do more with less
+Drill down
+Drink the Kool Aid
+Elephant in the room
+Engagement / engage them
+Feedback loop
+Get all your ducks in a row
+Heads up
+Hit the ground running
+It is what it is
+Key focus/priorities
+Limited bandwidth
+Look under the bonnet/hood
+Lots of moving parts
+Low-hanging fruit
+Move the goal posts
+Move the needle
+Moving forward/going forward
+On my plate/on my radar
+One-pager
+Open the kimono
+Out of the loop
+Par for the course
+Paradigm shift
+Park it
+Peel back the layers of the onion
+Ping
+Push the envelope
+Reach out
+Reinvent the wheel
+Run it up the flagpole
+Socialise
+Streamline
+Swamped
+SWAT Team
+Synergise
+Table the discussion
+Take it offline
+Tee it up
+Think outside the box
+Thought leader
+Throw under the bus
+Touch base
+Utilise
+When push comes to shove
