+++ my-app/src/components/DataDisplay.js
@@ -1,17 +1,33 @@
 import React from "react";
 import { useState, useEffect } from 'react';
 
-function DataDisplay(props) {
+export default function DataDisplay(){
+    const [randomWord, setRandomWord] = useState('');
 
+    async function fetchWords() {
+        const response = await fetch('words.txt');
+        const text = await response.text();
+        return text.trim().split('\n');
+    }
 
+    function getRandomWord(words) {
+        const randomIndex = Math.floor(Math.random() * words.length);
+        return words[randomIndex];
+    }
 
-    const {content} = props; 
-    
+    useEffect(() => {
+        async function updateRandomWord() {
+        const words = await fetchWords();
+        const newRandomWord = getRandomWord(words);
+        setRandomWord(newRandomWord);
+        }
+
+        updateRandomWord();
+    }, []);
+
     return (
-       
-    
-        <p>{content}</p>
-       
-    );
+      <>
+        {randomWord}
+      </>
+    ); 
 }
-export default DataDisplay;
